Remove exported resolver cache entry creation funcs.
NewOperatorForBundle and NewOperatorFromV1Alpha1CSV couple the
resolver cache package to the registry gRPC API and the
operators/v1alpha1 CSV API, respectively. They're effectively used
only by the resolver package, so they can be moved there and
unexported. The remaining calls to NewOperatorFromV1Alpha1CSV can be
extracted from core resolver logic into a Source implementation that
is responsible for the conversion from ClusterServiceVersion to
cache.Snapshot entry, and the remaining call to NewOperatorFromBundle
is used only by the Source implementation based on the registry gRPC
API.
