+++ pkg/controller/registry/resolver/cache/operators.go
@@ -1,7 +1,6 @@
 package cache
 
 import (
-	"encoding/json"
 	"fmt"
 	"hash/fnv"
 	"sort"
@@ -224,128 +223,6 @@ type Operator struct {
 	Bundle *api.Bundle
 }
 
-func NewOperatorFromBundle(bundle *api.Bundle, startingCSV string, sourceKey SourceKey, defaultChannel string) (*Operator, error) {
-	parsedVersion, err := semver.ParseTolerant(bundle.Version)
-	version := &parsedVersion
-	if err != nil {
-		version = nil
-	}
-	provided := APISet{}
-	for _, gvk := range bundle.ProvidedApis {
-		provided[opregistry.APIKey{Plural: gvk.Plural, Group: gvk.Group, Kind: gvk.Kind, Version: gvk.Version}] = struct{}{}
-	}
-	required := APISet{}
-	for _, gvk := range bundle.RequiredApis {
-		required[opregistry.APIKey{Plural: gvk.Plural, Group: gvk.Group, Kind: gvk.Kind, Version: gvk.Version}] = struct{}{}
-	}
-	sourceInfo := &OperatorSourceInfo{
-		Package:     bundle.PackageName,
-		Channel:     bundle.ChannelName,
-		StartingCSV: startingCSV,
-		Catalog:     sourceKey,
-	}
-	sourceInfo.DefaultChannel = sourceInfo.Channel == defaultChannel
-
-	// legacy support - if the api doesn't contain properties/dependencies, build them from required/provided apis
-	properties := bundle.Properties
-	if len(properties) == 0 {
-		properties, err = providedAPIsToProperties(provided)
-		if err != nil {
-			return nil, err
-		}
-	}
-	if len(bundle.Dependencies) > 0 {
-		ps, err := LegacyDependenciesToProperties(bundle.Dependencies)
-		if err != nil {
-			return nil, fmt.Errorf("failed to translate legacy dependencies to properties: %w", err)
-		}
-		properties = append(properties, ps...)
-	} else {
-		ps, err := RequiredAPIsToProperties(required)
-		if err != nil {
-			return nil, err
-		}
-		properties = append(properties, ps...)
-	}
-
-	o := &Operator{
-		Name:         bundle.CsvName,
-		Replaces:     bundle.Replaces,
-		Version:      version,
-		ProvidedAPIs: provided,
-		RequiredAPIs: required,
-		SourceInfo:   sourceInfo,
-		Properties:   properties,
-		Skips:        bundle.Skips,
-		BundlePath:   bundle.BundlePath,
-	}
-
-	if r, err := semver.ParseRange(bundle.SkipRange); err == nil {
-		o.SkipRange = r
-	}
-
-	if o.BundlePath == "" {
-		// This bundle's content is embedded within the Bundle
-		// proto message, not specified via image reference.
-		o.Bundle = bundle
-	}
-
-	return o, nil
-}
-
-func NewOperatorFromV1Alpha1CSV(csv *v1alpha1.ClusterServiceVersion) (*Operator, error) {
-	providedAPIs := EmptyAPISet()
-	for _, crdDef := range csv.Spec.CustomResourceDefinitions.Owned {
-		parts := strings.SplitN(crdDef.Name, ".", 2)
-		if len(parts) < 2 {
-			return nil, fmt.Errorf("error parsing crd name: %s", crdDef.Name)
-		}
-		providedAPIs[opregistry.APIKey{Plural: parts[0], Group: parts[1], Version: crdDef.Version, Kind: crdDef.Kind}] = struct{}{}
-	}
-	for _, api := range csv.Spec.APIServiceDefinitions.Owned {
-		providedAPIs[opregistry.APIKey{Group: api.Group, Version: api.Version, Kind: api.Kind, Plural: api.Name}] = struct{}{}
-	}
-
-	requiredAPIs := EmptyAPISet()
-	for _, crdDef := range csv.Spec.CustomResourceDefinitions.Required {
-		parts := strings.SplitN(crdDef.Name, ".", 2)
-		if len(parts) < 2 {
-			return nil, fmt.Errorf("error parsing crd name: %s", crdDef.Name)
-		}
-		requiredAPIs[opregistry.APIKey{Plural: parts[0], Group: parts[1], Version: crdDef.Version, Kind: crdDef.Kind}] = struct{}{}
-	}
-	for _, api := range csv.Spec.APIServiceDefinitions.Required {
-		requiredAPIs[opregistry.APIKey{Group: api.Group, Version: api.Version, Kind: api.Kind, Plural: api.Name}] = struct{}{}
-	}
-
-	properties, err := providedAPIsToProperties(providedAPIs)
-	if err != nil {
-		return nil, err
-	}
-	dependencies, err := RequiredAPIsToProperties(requiredAPIs)
-	if err != nil {
-		return nil, err
-	}
-	properties = append(properties, dependencies...)
-
-	return &Operator{
-		Name:         csv.GetName(),
-		Version:      &csv.Spec.Version.Version,
-		ProvidedAPIs: providedAPIs,
-		RequiredAPIs: requiredAPIs,
-		SourceInfo:   &ExistingOperator,
-		Properties:   properties,
-	}, nil
-}
-
-func (o *Operator) GetProvidedAPIs() APISet {
-	return o.ProvidedAPIs
-}
-
-func (o *Operator) GetRequiredAPIs() APISet {
-	return o.RequiredAPIs
-}
-
 func (o *Operator) Package() string {
 	if si := o.SourceInfo; si != nil {
 		return si.Package
@@ -358,200 +235,4 @@ func (o *Operator) Channel() string {
 		return si.Channel
 	}
 	return ""
-}
-
-func RequiredAPIsToProperties(apis APISet) (out []*api.Property, err error) {
-	if len(apis) == 0 {
-		return
-	}
-	out = make([]*api.Property, 0)
-	for a := range apis {
-		val, err := json.Marshal(struct {
-			Group   string `json:"group"`
-			Version string `json:"version"`
-			Kind    string `json:"kind"`
-		}{
-			Group:   a.Group,
-			Version: a.Version,
-			Kind:    a.Kind,
-		})
-		if err != nil {
-			return nil, err
-		}
-		out = append(out, &api.Property{
-			Type:  "olm.gvk.required",
-			Value: string(val),
-		})
-	}
-	if len(out) > 0 {
-		return
-	}
-	return nil, nil
-}
-
-func providedAPIsToProperties(apis APISet) (out []*api.Property, err error) {
-	out = make([]*api.Property, 0)
-	for a := range apis {
-		val, err := json.Marshal(opregistry.GVKProperty{
-			Group:   a.Group,
-			Version: a.Version,
-			Kind:    a.Kind,
-		})
-		if err != nil {
-			panic(err)
-		}
-		out = append(out, &api.Property{
-			Type:  opregistry.GVKType,
-			Value: string(val),
-		})
-	}
-	if len(out) > 0 {
-		return
-	}
-	return nil, nil
-}
-
-func LegacyDependenciesToProperties(dependencies []*api.Dependency) ([]*api.Property, error) {
-	var result []*api.Property
-	for _, dependency := range dependencies {
-		switch dependency.Type {
-		case "olm.gvk":
-			type gvk struct {
-				Group   string `json:"group"`
-				Version string `json:"version"`
-				Kind    string `json:"kind"`
-			}
-			var vfrom gvk
-			if err := json.Unmarshal([]byte(dependency.Value), &vfrom); err != nil {
-				return nil, fmt.Errorf("failed to unmarshal legacy 'olm.gvk' dependency: %w", err)
-			}
-			vto := gvk{
-				Group:   vfrom.Group,
-				Version: vfrom.Version,
-				Kind:    vfrom.Kind,
-			}
-			vb, err := json.Marshal(&vto)
-			if err != nil {
-				return nil, fmt.Errorf("unexpected error marshaling generated 'olm.package.required' property: %w", err)
-			}
-			result = append(result, &api.Property{
-				Type:  "olm.gvk.required",
-				Value: string(vb),
-			})
-		case "olm.package":
-			var vfrom struct {
-				PackageName  string `json:"packageName"`
-				VersionRange string `json:"version"`
-			}
-			if err := json.Unmarshal([]byte(dependency.Value), &vfrom); err != nil {
-				return nil, fmt.Errorf("failed to unmarshal legacy 'olm.package' dependency: %w", err)
-			}
-			vto := struct {
-				PackageName  string `json:"packageName"`
-				VersionRange string `json:"versionRange"`
-			}{
-				PackageName:  vfrom.PackageName,
-				VersionRange: vfrom.VersionRange,
-			}
-			vb, err := json.Marshal(&vto)
-			if err != nil {
-				return nil, fmt.Errorf("unexpected error marshaling generated 'olm.package.required' property: %w", err)
-			}
-			result = append(result, &api.Property{
-				Type:  "olm.package.required",
-				Value: string(vb),
-			})
-		case "olm.label":
-			result = append(result, &api.Property{
-				Type:  "olm.label.required",
-				Value: dependency.Value,
-			})
-		}
-	}
-	return result, nil
-}
-
-func APISetToProperties(crds, apis APISet, deprecated bool) (out []*api.Property) {
-	out = make([]*api.Property, 0)
-	for a := range crds {
-		val, err := json.Marshal(opregistry.GVKProperty{
-			Group:   a.Group,
-			Kind:    a.Kind,
-			Version: a.Version,
-		})
-		if err != nil {
-			panic(err)
-		}
-		out = append(out, &api.Property{
-			Type:  opregistry.GVKType,
-			Value: string(val),
-		})
-	}
-	for a := range apis {
-		val, err := json.Marshal(opregistry.GVKProperty{
-			Group:   a.Group,
-			Kind:    a.Kind,
-			Version: a.Version,
-		})
-		if err != nil {
-			panic(err)
-		}
-		out = append(out, &api.Property{
-			Type:  opregistry.GVKType,
-			Value: string(val),
-		})
-	}
-	if deprecated {
-		val, err := json.Marshal(opregistry.DeprecatedProperty{})
-		if err != nil {
-			panic(err)
-		}
-		out = append(out, &api.Property{
-			Type:  opregistry.DeprecatedType,
-			Value: string(val),
-		})
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return
-}
-
-func APISetToDependencies(crds, apis APISet) (out []*api.Dependency) {
-	if len(crds)+len(apis) == 0 {
-		return nil
-	}
-	out = make([]*api.Dependency, 0)
-	for a := range crds {
-		val, err := json.Marshal(opregistry.GVKDependency{
-			Group:   a.Group,
-			Kind:    a.Kind,
-			Version: a.Version,
-		})
-		if err != nil {
-			panic(err)
-		}
-		out = append(out, &api.Dependency{
-			Type:  opregistry.GVKType,
-			Value: string(val),
-		})
-	}
-	for a := range apis {
-		val, err := json.Marshal(opregistry.GVKDependency{
-			Group:   a.Group,
-			Kind:    a.Kind,
-			Version: a.Version,
-		})
-		if err != nil {
-			panic(err)
-		}
-		out = append(out, &api.Dependency{
-			Type:  opregistry.GVKType,
-			Value: string(val),
-		})
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return
 }
